@@ -1,5 +1,5 @@
 ﻿// <auto-generated />
-using DanceCoolDataAccessLogic.EfStructures.Context;
+using DataAccessLogic.EfStructures.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
